mlx4_core: alias_GUID.c, mcg.c, sysfs.c Fix printf formats.

+++ drivers/infiniband/hw/mlx4/alias_GUID.c
@@ -82,7 +82,8 @@ void mlx4_ib_update_cache_on_guid_change(struct mlx4_ib_dev *dev, int block_num,
 	guid_indexes = be64_to_cpu((__force __be64) dev->sriov.alias_guid.
 				   ports_guid[port_num - 1].
 				   all_rec_per_port[block_num].guid_indexes);
-	pr_debug("port: %d, guid_indexes: 0x%llx\n", port_num, guid_indexes);
+	pr_debug("port: %d, guid_indexes: 0x%llx\n", port_num,
+	    (long long)guid_indexes);
 
 	for (i = 0; i < NUM_ALIAS_GUID_IN_REC; i++) {
 		/* The location of the specific index starts from bit number 4
@@ -146,7 +147,8 @@ void mlx4_ib_notify_slaves_on_guid_change(struct mlx4_ib_dev *dev,
 	guid_indexes = be64_to_cpu((__force __be64) dev->sriov.alias_guid.
 				   ports_guid[port_num - 1].
 				   all_rec_per_port[block_num].guid_indexes);
-	pr_debug("port: %d, guid_indexes: 0x%llx\n", port_num, guid_indexes);
+	pr_debug("port: %d, guid_indexes: 0x%llx\n", port_num,
+	    (long long)guid_indexes);
 
 	/*calculate the slaves and notify them*/
 	for (i = 0; i < NUM_ALIAS_GUID_IN_REC; i++) {
@@ -290,7 +292,7 @@ static void aliasguid_query_handler(int status,
 					     "new val(0x%llx) was kept\n",
 					      __func__, i,
 					     guid_rec->block_num,
-					     be64_to_cpu(*(__be64 *) &
+					     (long long)be64_to_cpu(*(__be64 *) &
 							 rec->all_recs[i * GUID_REC_SIZE]));
 			} else {
 				memcpy(&rec->all_recs[i * GUID_REC_SIZE],
