+++ drivers/infiniband/hw/mlx4/sysfs.c
@@ -57,7 +57,7 @@ static ssize_t show_admin_alias_guid(struct device *dev,
 	guid_index_in_rec = mlx4_ib_iov_dentry->entry_num % 8 ;
 
 	return sprintf(buf, "%llx\n",
-		       be64_to_cpu(*(__be64 *)&mdev->sriov.alias_guid.
+		       (long long)be64_to_cpu(*(__be64 *)&mdev->sriov.alias_guid.
 				   ports_guid[port->num - 1].
 				   all_rec_per_port[record_num].
 				   all_recs[8 * guid_index_in_rec]));
